+++ utils/src/main/lombok/de/larssh/utils/time/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Utilities for date and time
+ */
+@de.larssh.utils.annotations.NonNullByDefault
+package de.larssh.utils.time;